refactor(rust): satisfy import-linter contracts for the façade
Two layering violations in rs_lint.py, both fixed by moving code to the
layer that owns it (no ignore_imports holes):

- core.rules must not import core.linter: apply_source_fixes,
  facade_violations, _native_apply_fixes and facade_fix_loop move
  verbatim to a new core/linter/rs_fix.py — applying fixes needs the
  linter's fix/patch machinery, so the loop belongs in the linter layer.
  rs_lint.py keeps the RsSegment façade and FACADE_SAFE_RULES.
- core must not import sqlfluff.utils: RsSegment now exposes raw
  reflow_depth_data() scalars and DepthMap.from_parent assembles the
  DepthInfo objects itself.
